New sliders working
