+++ package.json
@@ -15,6 +15,8 @@
     "@trpc/react": "^9.20.3",
     "@trpc/server": "^9.20.3",
     "@types/lodash-es": "^4.17.5",
+    "@types/node": "^14.14.6",
+    "@types/react": "^16.9.56",
     "@types/uuid": "^8.3.0",
     "axios": "^0.21.1",
     "classnames": "^2.3.1",
@@ -23,15 +25,15 @@
     "mongodb": "^4.2.2",
     "next": "^11.1.3",
     "react": "^17.0.2",
+    "react-aria": "^3.19.0",
     "react-dom": "^17.0.2",
     "react-ga": "^3.3.0",
     "react-query": "^3.34.16",
     "react-spring-bottom-sheet": "^3.4.0",
+    "react-stately": "^3.17.0",
     "react-window": "^1.8.6",
     "zod": "^3.14.2",
-    "zustand": "^3.6.7",
-    "@types/node": "^14.14.6",
-    "@types/react": "^16.9.56"
+    "zustand": "^3.6.7"
   },
   "devDependencies": {
     "@types/lodash": "^4.14.178",
